requirements.txt
django-cleanup
django-allauth
django-htmx